Add basic broker API
This builds the service broker HTTP handler and tests it serves a
healthcheck endpoint. The intention is to use this set of tests for
the implementation of the Open Service Broker API. The BrokerClient
should eventually be split out into a separate library, as it will be
useful for testing other brokers.