Divide config into API, Catalog, and ProviderData
We unmarshal the same config three times to make a clear distinction
between config for the broker API, the catalog to be returned to the
broker client (usually Cloud Controller), and the data which is specific
to the provider (an AWS client for AWS services for example).

Using the brokerapi library means when we try to unmarshal configuration
files into a catalog the upstream datatypes do not have anywhere for us
to add extra configuration details for a service or plan which is unique
to a given provider. For example, if the plan was for an AWS service I
may want to add a database instance class to the plan details. The aim
will be to use a provider interface which takes responsibility for
unmarshalling the raw JSON message into its down data structures.