Add port and log level to API config
We want to be able to pass these values into the config file to make
it easier to change a broker's port and logging level.

We add a helper function to convert text values into lager.LogLevel
values, as these are what will be used to create the broker's logger.