Add an Aiven client
There were few existing Golang clients for Aiven, so we decided to write
our own. The reason for not using the most popular open source client
was it wasn't tested or written in a way which made testing easy. Given
the easy to use nature of the Aiven API it did not take long to write
our own.