Add an Aiven provider
The upstream library we built this broker upon defines a
`ServiceProvider` interface to implement. This provides the necessary
implementation and the data structures to unmarshal the configuration
file into. This is deliberately kept separate to the configuration of
the broker itself, so some duplication of tests and looking up plans and
services may be seen.