Update Aiven client
To support `Bind` operations we needed to create the service user and
combine it with getting service connection details.

For `LastOperation` we need to get both the state and the update time
for a service. This is because a service will report its state as
'RUNNING' immediately after a request to change plan, even though the
update has not completed. This would trick brokers into thinking the
update has finished.  See the following commits for the implementation
of this in the broker/provider.