Implement Bind/Unbind/Update
For `LastOperation` we need to get both the state and the update time
for a service. This is because a service will report its state as
'RUNNING' immediately after a request to change plan, even though the
update has not completed. This would trick brokers into thinking the
update has finished.