Add integration tests for bind/unbind and update
There are numerous race conditions with Aiven's API which affect how
we must do the integration test:

1. The hostname for a service does not resolve immediately. This means
trying to bind and connect straight after creation often fails.

2. After a service is created you cannot update it until a backup has
been made by Aiven. We do not expect users to update an instance seconds
or minutes after creating one, so this is something we only have to work
around in tests.

3. When deleting a user the API responds immediately, but the user is
still able to connect for a few seconds. This causes failures when
trying to assert the unbinding was successful.