Stop hashing the instanceId to generate the instance name
The instance name was previously limited to 16 characters, so we had to
hash the instanceID (a GUID) using crc32 to get something that would
fit.

Since then Aiven have relaxed this limit and it's now possible to
use instance names up to 64 characters. This therefore removes the crc32
hashing, and merely appends the instanceID to the prefix.

Now that the instance name generation is vastly simplified, there's no
longer a need for it to be an exported function as the integration tests
don't need to make use of it.