Stop exporting provider.ProviderStatesMapping
This is an internal function, and therefore shouldn't be exported from
the package.

In order to retain the tests for this, I've used the pattern of having a
separate test file for the internals of a package. This means that we
can test private functions while still keeping the rest of the tests
restricted to the public interface to the package.