Add a CF manifest.yml
We're currently providing in our deployment pipeline from a separate
repo. Given most of these values are specific to the codebase itself
instead of the deployment, it makes more sense to include it here. This
will also make this repo more useful for 3rd parties who want to use it.