Return specific error if instance already deleted
The upstream brokerapi library defines a specific error type to return
if a service instance does not exist. This type is used to return a 410
response code when trying to delete an instance that does not exist
(typically because it has already been deleted).

Aiven returns a 404 if you try to delete an instance that does not
exist, so we refactor our provider to translate this into the error type
required by the brokerapi library.