Service update include user config and errors
When performing the instance update we must also include
the user_config struct with specifies the elasticsearch
version and IP filter.

Including the ES version would allow do version upgrades, and
the IP filter is needed to avoid potentially wiping the
existing values.

Additionally, Aiven would return an error if the udate is not
possible: major version downgrade, clustered to unclustered changes,
etc. This error is returned as 400 status and a json with the
human readable error.

We capture this error and return it
to the client as a 422 StatusUnprocessableEntity, as specified in
the broker api spec[1]. For any other error we also return the
body of the response, as it might be useful when troubleshooting
issues.

Tests have been updated to reflect all these changes.

[1] https://github.com/openservicebrokerapi/servicebroker/blob/v2.12/spec.md#response-3