Ensure tests cleanup instance on failure.
The IP whitelisting test was cleaning up with a defer, but the main
lifecycle one wasn't. This moves the cleanup into an AfterEach so that
it gets run for all the tests. The unbind is unnecessary, as a
deprovision will delete the instance, and there's no state kept anywhere
other than in the instance.