Add a poll of the ES endpoint in a Bind call
We've seen issues where the credentials returned by the Aiven API don't
work for a short period after creation - the ES endpoint returns 401
errors. This has caused things like our acceptance tests to fail.

This adds a polling loop to the bind call to wait until the credentials
successfully connect. Becuase this is only a best-effort workaround, if
the polling times out, it returns anyway.

With this in place, there's no longer a need to have the polling in the
integration tests, so this is also removed.