Improve resiliency of Unbind to errors from Aiven
In recent support ticket #3769643, a tenant reported that unbinding
their app from an Elasticsearch was failing:

    Service broker error: Error deleting service user: 403 status code
    returned from Aiven: '{"errors":[{"message":"Service user
    'GUIDGUID-GUID-GUID-GUID-GUIDGUIDGUID' does not exist

(Where GUID is a placeholder for the real GUID involved.)

I investigated and found that on the first attempt to unbind, Aiven
returned a 502 Bad Gateway error when the broker attempted to delete
the binding's Elasticsearch user. It seems that Aiven did delete the
user despite that error response.

This commit has the broker detect when a user does not exist, such as
when it has already been deleted. This should prevent this problem
from recurring and help maintain our tenants' confidence in us.