include influxdb db name in binding credentials
includes the name of the provided database in the binding credentials so
that we have all the information required to start reading/writing to
the database.

the influxdb username/password/host/port provided in the credentials are
not enough to connect and start using the database ... you also need to
either create a database or use an existing one.

the provided credentials do not have permission to create databases, so
we need to know the name of the database that is provided for us.

the database name "defaultdb" appears currently hardcoded in the
prometheus-specific configuration credentials, but it isn't fun to parse
it from there or to hardcoded it anywhere incase it changes.