Add timestamp change
The pipelines are currently broken after removing TLWR from the
"trusted" people. This is unfortunate, when it comes to repos that
require signed commits, but also aren't maintained very often.

Concourse is likely to ignore an empty signed commit. Meaning there
needs to be a codebase change for the pipelines to kick through.

Adding the timestamp to the readmem which is essentially a HTML comment,
will satisfy that problem, and allow us to bump the codebase without
finding perhaps unknown bugs.