Set the broker name in integration tests
The broker name is added as a tag to the service. Previously we have
just set this to 'test-broker' when running integration tests. Instead,
set it to something that identifes the test run.

The name is dependent on the environment the tests are running in:
  locally:   'integration_local_<commit sha>'
  concourse: 'integration_concourse_pr<pr number>_<commit sha>'

This should better enable us to both identify dangling services if tests
fail, and also to more easily identify services created by a specific
test run.